% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{article_locate}
\alias{article_locate}
\title{Locate articles matching a pattern}
\usage{
article_locate(pattern, collection = "_posts")
}
\arguments{
\item{pattern}{Regular expression}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Character vector of paths
}
\description{
Locate articles matching a pattern
}
