% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{article_named}
\alias{article_named}
\title{Locate articles matching a pattern}
\usage{
article_named(pattern, collection = "posts")
}
\arguments{
\item{pattern}{Regular expression}

\item{collection}{The collection the article belongs to (default = "posts")}
}
\value{
Character vector of paths
}
\description{
Locate articles matching a pattern
}
