% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{article_sitrep}
\alias{article_sitrep}
\title{Situation report for an article}
\usage{
article_sitrep(dir, collection = "posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "posts")}
}
\value{
Invisibly returns a list containing the paths
}
\description{
Situation report for an article
}
\details{
Provides a minimal "situation report" for a post, listing the
folders associated with the post. Later versions will check renv status, I
hope.
}
