% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{articles_matching}
\alias{articles_matching}
\title{Returns articles matching a pattern}
\usage{
articles_matching(pattern, collection = "_posts")
}
\arguments{
\item{pattern}{Regular expression}

\item{collection}{Folder to look in}
}
\value{
Character vector of paths
}
\description{
Returns articles matching a pattern
}
