% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{insert_appendix}
\alias{insert_appendix}
\title{Inserts the default refinery appendix}
\usage{
insert_appendix(repo_spec, dir, collection = "posts")
}
\arguments{
\item{repo_spec}{Repository specification, formatted as "user/repo_name"}

\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "posts")}
}
\value{
A shiny.tag object
}
\description{
Inserts the default refinery appendix
}
