% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{insert_netlify_redirect}
\alias{insert_netlify_redirect}
\title{Setup redirections for netlify deployment}
\usage{
insert_netlify_redirect(slug, date, collection = "posts", publish = "_site")
}
\arguments{
\item{slug}{The slug for the post}

\item{date}{The date for the post}

\item{collection}{The collection (default = "posts")}

\item{publish}{The publish directory (default = "_site")}
}
\value{
???
}
\description{
Setup redirections for netlify deployment
}
