% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{insert_source}
\alias{insert_source}
\title{Inserts a link to the source code}
\usage{
insert_source(
  repo_spec,
  dir,
  collection = "_posts",
  branch = "master",
  host = "https://github.com",
  text = "source code"
)
}
\arguments{
\item{repo_spec}{Repository specification, formatted as "user/repo_name"}

\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}

\item{branch}{Name of the branch to look at}

\item{host}{Host URL}

\item{text}{Text to display in the link}
}
\value{
A "shiny.tag" object
}
\description{
Inserts a link to the source code
}
