% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{post_path}
\alias{post_path}
\title{Locate a distill article}
\usage{
post_path(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Character string with the path
}
\description{
Locate a distill article
}
