% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{record_lockfile}
\alias{record_lockfile}
\title{Records a link to the lock file}
\usage{
record_lockfile(
  repo_spec,
  dir,
  collection = "_posts",
  branch = "master",
  host = "https://github.com"
)
}
\arguments{
\item{repo_spec}{Repository specification, formatted as "user/repo_name"}

\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}

\item{branch}{Name of the branch to look at}

\item{host}{Host URL}
}
\value{
A "shiny.tag" object
}
\description{
Records a link to the lock file
}
