% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{record_timestamp}
\alias{record_timestamp}
\title{Records a time stamp to the R markdown output}
\usage{
record_timestamp(tzone = Sys.timezone())
}
\arguments{
\item{tzone}{Character vector specifying the time zone}
}
\value{
A "shiny.tag" object
}
\description{
Records a time stamp to the R markdown output
}
