% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_activate}
\alias{renv_activate}
\title{Activate the R environment associated with a distll article}
\usage{
renv_activate(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Invisibly returns the renv project directory
}
\description{
Activate the R environment associated with a distll article
}
