% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_delete}
\alias{renv_delete}
\title{Delete the renv set up associated with a post}
\usage{
renv_delete(dir, collection = "posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "posts")}
}
\value{
??
}
\description{
Delete the renv set up associated with a post
}
