% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_library}
\alias{renv_library}
\title{Locate the library for the post}
\usage{
renv_library(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
path
}
\description{
Locate the library for the post
}
