% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_load}
\alias{renv_load}
\title{Load the renv library}
\usage{
renv_load(dir, collection = "posts", ...)
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "posts")}

\item{...}{Arguments to be passed to renv::load()}
}
\value{
???
}
\description{
Load the renv library
}
