% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_lockfile}
\alias{renv_lockfile}
\title{Locate the renv lockfile associated with a distill article}
\usage{
renv_lockfile(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Character string with the path
}
\description{
Locate the renv lockfile associated with a distill article
}
\details{
The renv.lock file is stored in the post folder
}
