% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_new}
\alias{renv_new}
\title{Create a minimal R environment for a post}
\usage{
renv_new(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
??
}
\description{
Create a minimal R environment for a post
}
\details{
Creates the renv library and installs renv, distill and refinery
}
