% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_snapshot}
\alias{renv_snapshot}
\title{Create a snapshot of the R environment for the post}
\usage{
renv_snapshot(dir, collection = "_posts", ...)
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}

\item{...}{Arguments to be passed to renv::snapshot()}
}
\value{
???
}
\description{
Create a snapshot of the R environment for the post
}
