% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_start}
\alias{renv_start}
\title{Create the renv setup}
\usage{
renv_start(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
??
}
\description{
Create the renv setup
}
