% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_use}
\alias{renv_use}
\title{Use the R environment associated with a distill article}
\usage{
renv_use(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Invisibly returns the renv project directory
}
\description{
Use the R environment associated with a distill article
}
