% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendices.R
\name{timestamp}
\alias{timestamp}
\title{Writes a time stamp to the R markdown output}
\usage{
timestamp(tzone = Sys.timezone())
}
\arguments{
\item{tzone}{Character vector specifying the time zone}
}
\value{
A "shiny.tag" object
}
\description{
Writes a time stamp to the R markdown output
}
