% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_article_template}
\alias{use_article_template}
\title{Create new article from existing template}
\usage{
use_article_template(
  template,
  slug,
  date = NULL,
  collection = "posts",
  slug_replace = "INSERT-SLUG-HERE",
  date_replace = "INSERT-DATE-HERE",
  renv_new = TRUE,
  open = interactive()
)
}
\arguments{
\item{template}{Path to template file}

\item{slug}{Semantic slug for the article (in-kebab-case)}

\item{date}{Date for the article (in YYYY-MM-DD)}

\item{collection}{Collection to contain the article (default = "posts")}

\item{slug_replace}{Pattern to be replaced with slug}

\item{date_replace}{Pattern to be replaced with date}

\item{renv_new}{Initialise a new R environment?}

\item{open}{Open the article in RStudio?}
}
\value{
???
}
\description{
Create new article from existing template
}
