% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{use_post_template}
\alias{use_post_template}
\title{Create new post from existing template}
\usage{
use_post_template(
  template,
  slug,
  date = NULL,
  collection = "_posts",
  slug_replace = "INSERT-SLUG-HERE",
  date_replace = "INSERT-DATE-HERE",
  renv_new = TRUE
)
}
\arguments{
\item{template}{Path to template file}

\item{slug}{Semantic slug for the post (in-kebab-case)}

\item{date}{Date for the post (in YYYY-MM-DD)}

\item{collection}{Collection to contain the post (default = "_posts")}

\item{slug_replace}{Pattern to be replaced with slug}

\item{date_replace}{Pattern to be replaced with date}

\item{renv_new}{Initialise a new R environment?}
}
\value{
???
}
\description{
Create new post from existing template
}
