% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_path}
\alias{renv_path}
\title{Locate the renv project environment for a distill article}
\usage{
renv_path(dir, collection = "_posts")
}
\arguments{
\item{dir}{The folder in which the article is located}

\item{collection}{The collection the article belongs to (default = "_posts")}
}
\value{
Character string with the path
}
\description{
Locate the renv project environment for a distill article
}
